Added calls to searchStorageAPI (interface) for updates on relative stuff on DAO Implementations

+++ back_end/src/main/java/back/data/jdbc/TransactionRowMapper.java
@@ -1,6 +1,5 @@
 package back.data.jdbc;
 
-import back.model.Image;
 import back.model.Transaction;
 import back.util.DateHandler;
 import org.springframework.jdbc.core.RowMapper;
@@ -14,7 +13,7 @@ public class TransactionRowMapper implements RowMapper<Transaction> {
     public Transaction mapRow(ResultSet rs, int rowNum) throws SQLException {
         int id = rs.getInt("id");
         int roomId = rs.getInt("room_id");
-        int visitorId = rs.getInt("visitor_id");
+        long visitorId = rs.getLong("visitor_id");
         double cost = rs.getDouble("cost");
         String startDate = DateHandler.SQLDateToFrontDate(rs.getString("start_date"));
         String endDate = DateHandler.SQLDateToFrontDate(rs.getString("end_date"));
